requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.35.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
instagrapi==2.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==2.1.5
migrate==0.3.8
narwhals==1.36.0
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.75.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
playwright==1.51.0
pocketsphinx==5.0.3
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
pymongo==4.8.0
pyparsing==3.2.3
PyPasser==0.0.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
selenium==4.23.1
setuptools==72.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.8.1
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
trio==0.26.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.44.0
wsproto==1.2.0
